requirements.txt
# Core dependencies
google-generativeai>=0.8.5
python-dotenv>=1.0.0

# Streamlit and UI dependencies
streamlit>=1.52.0
plotly>=6.5.0
streamlit-option-menu>=0.4.0

# Data handling
pandas>=2.3.0
numpy>=1.23.0

# Other utilities
requests>=2.32.0